add the PopoverSerivce and breakup interfaces
