+++ client-course-schedulizer/src/components/reuseables/AddSectionPopover/index.ts
@@ -1,1 +1,2 @@
 export * from "./AddSectionPopover";
+export * from "./AddSectionPopoverService";
