+++ client-course-schedulizer/src/utilities/interfaces/dataInterfaces/dayInterfaces.ts
@@ -1,0 +1,19 @@
+/* eslint-disable typescript-sort-keys/string-enum */
+
+export enum Day {
+  Monday = "M",
+  Tuesday = "T",
+  Wednesday = "W",
+  Thursday = "Th",
+  Friday = "F",
+  Saturday = "S",
+  Sunday = "Su",
+}
+
+export enum Weekday {
+  Monday = Day.Monday,
+  Tuesday = Day.Tuesday,
+  Wednesday = Day.Wednesday,
+  Thursday = Day.Thursday,
+  Friday = Day.Friday,
+}
