+++ client-course-schedulizer/src/utilities/interfaces/dataInterfaces/index.ts
@@ -1,0 +1,3 @@
+export * from "./dayInterfaces";
+export * from "./scheduleInterfaces";
+export * from "./termInterfaces";
