+++ client-course-schedulizer/src/utilities/interfaces/dataInterfaces/scheduleInterfaces.ts
@@ -1,56 +1,5 @@
 /* eslint-disable typescript-sort-keys/string-enum */
-export enum Term {
-  Fall = "FA",
-  Interim = "IN", // TODO: Remove?
-  Spring = "SP",
-  Summer = "SU", // TODO: Is this a thing?
-}
-
-export enum SemesterLength {
-  Full = "Full",
-  HalfFirst = "First",
-  HalfSecond = "Second",
-  IntensiveA = "A",
-  IntensiveB = "B",
-  IntensiveC = "C",
-  IntensiveD = "D",
-}
-
-export enum Day {
-  Monday = "M",
-  Tuesday = "T",
-  Wednesday = "W",
-  Thursday = "Th",
-  Friday = "F",
-  Saturday = "S",
-  Sunday = "Su",
-}
-
-export enum Weekday {
-  Monday = Day.Monday,
-  Tuesday = Day.Tuesday,
-  Wednesday = Day.Wednesday,
-  Thursday = Day.Thursday,
-  Friday = Day.Friday,
-}
-
-export enum Half {
-  First = SemesterLength.HalfFirst,
-  Second = SemesterLength.HalfSecond,
-}
-
-export enum Intensive {
-  A = SemesterLength.IntensiveA,
-  B = SemesterLength.IntensiveB,
-  C = SemesterLength.IntensiveC,
-  D = SemesterLength.IntensiveD,
-}
-
-export enum SemesterLengthOption {
-  FullSemester = "Full",
-  HalfSemester = "Half",
-  IntensiveSemester = "Intensive",
-}
+import { Day, SemesterLength, Term } from ".";
 
 export interface Location {
   building: string;
